Remove Moment.js date parsing
The date parsing code tries to guess the date format and often gets it
wrong. YNAB parses multiple date formats and presents the date format
options to the user in the web UI. I think it's better to just accept
the raw date and let YNAB handle the parsing later.

This commit removes extra parsing for the date field.

Resolves #9.